A lot of changes internally, including: ntegrating "gameclient" into the main class, turning it into a singleton improving the PlayerManager (AKA formerly controller manager) Player and VesselController (aka formerly control tracker)  to facilitate support for more controllers. Setting my IDE up to ignore the state machine related content.
but the big noticeable change from a user perspective is that you can now switch between controller and keyboard by hitting 1 and 2 respectively (on the row with the special characters, not the numpad.), if you attempt to force connect to a controller without a controller connected, nothing will happen.

Also a number of small bug fixes.

new zip uploaded reflecting these changes.